Enable requesting multiple types of evidence and adding multipart form
